Adjust type conversion for BrowserFunctions
Convert return types of BrowserFunctions according to the contract
defined by SWT. Do not use JsonUtil, which converts more types than
permitted by SWT.

Convert return type right after calling the function and throw an SWT
error in case of invalid return type.

Update error messages copy from SWT.

+++ tests/org.eclipse.rap.rwt.test/src/org/eclipse/swt/internal/browser/browserkit/BrowserLCA_Test.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2008, 2013 Innoopract Informationssysteme GmbH and others.
+ * Copyright (c) 2008, 2014 Innoopract Informationssysteme GmbH and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -366,7 +366,7 @@ public class BrowserLCA_Test {
       .add( JsonValue.NULL )
       .add( new JsonArray().add( "a string" ).add( false ) )
       .add( "hi" )
-      .add( 0.6666667f )
+      .add( Float.valueOf( 0.6666667f ).doubleValue() )
       .add( 12l );
     JsonArray expected = new JsonArray()
       .add( "func" )
